# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NOVA Ride API",
    "version": "1.0.0",
    "description": "جرد مسارات مُولّد آليًا من المتحكّمات. البادئة /api؛ كل المسارات متاحة أيضًا تحت /api/v1. لا يتضمّن هذا الجرد مخطّطات الطلب/الاستجابة (تُثرى لاحقًا)."
  },
  "servers": [
    {
      "url": "{baseUrl}",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:4000"
        }
      }
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/admin/fare-offers": {
      "get": {
        "tags": [
          "fare-quotes"
        ],
        "operationId": "fare-quotes_list",
        "summary": "GET /api/admin/fare-offers",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: pricing.manage"
      }
    },
    "/api/admin/fare-offers/:id": {
      "get": {
        "tags": [
          "fare-quotes"
        ],
        "operationId": "fare-quotes_getOne",
        "summary": "GET /api/admin/fare-offers/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: pricing.manage"
      }
    },
    "/api/admin/fare-quotes": {
      "get": {
        "tags": [
          "fare-quotes"
        ],
        "operationId": "fare-quotes_list",
        "summary": "GET /api/admin/fare-quotes",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: pricing.manage"
      }
    },
    "/api/admin/fare-quotes/:id": {
      "get": {
        "tags": [
          "fare-quotes"
        ],
        "operationId": "fare-quotes_getOne",
        "summary": "GET /api/admin/fare-quotes/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: pricing.manage"
      }
    },
    "/api/admin/fare-quotes/simulate": {
      "post": {
        "tags": [
          "fare-quotes"
        ],
        "operationId": "fare-quotes_simulate",
        "summary": "POST /api/admin/fare-quotes/simulate",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: pricing.manage"
      }
    },
    "/api/admin/geo/provider": {
      "get": {
        "tags": [
          "geo"
        ],
        "operationId": "geo_getConfig",
        "summary": "GET /api/admin/geo/provider",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      },
      "put": {
        "tags": [
          "geo"
        ],
        "operationId": "geo_updateConfig",
        "summary": "PUT /api/admin/geo/provider",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      }
    },
...
    "/api/documents/:id/review": {
      "patch": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_review",
        "summary": "PATCH /api/documents/:id/review",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: drivers.documents, drivers.manage"
      }
    },
...
    "/api/drivers": {
      "get": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_findAll",
        "summary": "GET /api/drivers",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: drivers.read, drivers.manage"
      }
    },
...
    "/api/drivers/:id": {
      "get": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_findOne",
        "summary": "GET /api/drivers/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: drivers.read, drivers.manage"
      }
    },
    "/api/drivers/:id/approve": {
      "patch": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_approve",
        "summary": "PATCH /api/drivers/:id/approve",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: drivers.manage"
      }
    },
    "/api/drivers/:id/ban": {
      "patch": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_ban",
        "summary": "PATCH /api/drivers/:id/ban",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: drivers.manage"
      }
    },
    "/api/drivers/:id/reject": {
      "patch": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_reject",
        "summary": "PATCH /api/drivers/:id/reject",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: drivers.manage"
      }
    },
    "/api/drivers/:id/suspend": {
      "patch": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_suspend",
        "summary": "PATCH /api/drivers/:id/suspend",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: drivers.manage"
      }
    },
    "/api/drivers/documents/:docId/review": {
      "patch": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_reviewDocument",
        "summary": "PATCH /api/drivers/documents/:docId/review",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: drivers.documents, drivers.manage"
      }
    },
    "/api/drivers/sanctions/:id/lift": {
      "patch": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_liftSuspension",
        "summary": "PATCH /api/drivers/sanctions/:id/lift",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: drivers.manage"
      }
    },
    "/api/drivers/sanctions/config": {
      "get": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_sanctionsConfig",
        "summary": "GET /api/drivers/sanctions/config",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: drivers.read, drivers.manage"
      }
    },
    "/api/drivers/sanctions/log": {
      "get": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_sanctionsLog",
        "summary": "GET /api/drivers/sanctions/log",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: drivers.read, drivers.manage"
      }
    },
    "/api/drivers/sanctions/suspended": {
      "get": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_suspendedDrivers",
        "summary": "GET /api/drivers/sanctions/suspended",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: drivers.read, drivers.manage"
      }
    },
...
    "/api/fare-quotes/:id": {
      "get": {
        "tags": [
          "fare-quotes"
        ],
        "operationId": "fare-quotes_getOne",
        "summary": "GET /api/fare-quotes/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/geo/autocomplete": {
      "get": {
        "tags": [
          "geo"
        ],
        "operationId": "geo_autocomplete",
        "summary": "GET /api/geo/autocomplete",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/geo/directions": {
      "post": {
        "tags": [
          "geo"
        ],
        "operationId": "geo_directions",
        "summary": "POST /api/geo/directions",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/geo/geocode": {
      "get": {
        "tags": [
          "geo"
        ],
        "operationId": "geo_geocode",
        "summary": "GET /api/geo/geocode",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/geo/places": {
      "get": {
        "tags": [
          "geo"
        ],
        "operationId": "geo_list",
        "summary": "GET /api/geo/places",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "geo"
        ],
        "operationId": "geo_create",
        "summary": "POST /api/geo/places",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/geo/places/:id": {
      "delete": {
        "tags": [
          "geo"
        ],
        "operationId": "geo_remove",
        "summary": "DELETE /api/geo/places/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "geo"
        ],
        "operationId": "geo_update",
        "summary": "PATCH /api/geo/places/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/geo/places/recent": {
      "post": {
        "tags": [
          "geo"
        ],
        "operationId": "geo_recordRecent",
        "summary": "POST /api/geo/places/recent",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/geo/reverse": {
      "get": {
        "tags": [
          "geo"
        ],
        "operationId": "geo_reverse",
        "summary": "GET /api/geo/reverse",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}